requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai>=1.35.0
pymongo
psycopg2-binary
sqlglot
sentence-transformers
torch
transformers
numpy